send userLeave on user eject
